Added supports for topics
The semaphore is incremented once for each receiver that gets a message when using a topic, rather than once for each message sent. Before the lock was released as soon as just one of the receivers acknoweldged the message, but we only want it released once all receivers had acknowledged a message.